Replace `SvgDoc` with `Document[Svg]` via `Loadable`
Drops the bespoke `SvgDoc` case class in favour of turbulence's existing
`Document[content <: Documentary]` wrapper, which is precisely the
concept being modelled: a parsed payload (`Svg`) plus envelope metadata
(`Encoding` from the XML declaration). This is what `Loadable` was
designed for, and the previous `Svg is Aggregable by Text` /
`SvgDoc is Aggregable by Text` split was reinventing the abstraction.

- `Svg` now extends `Documentary` with `type Metadata = Encoding`.
- `Svg is Loadable by Text` parses the optional `<?xml … encoding=…?>`
  declaration off the front and returns `Document[Svg]`. The XML-prologue
  string handling that used to live in `SvgDoc.aggregable` moves here,
  but the encapsulation now matches the abstraction's intent.
- `Document[Svg] is Showable` produces the round-trip text
  (`<?xml?><svg…/>`) for callers that want a single `.show: Text`. The
  given is parameterised as `[doc <: Document[Svg]]` so it matches the
  path-refined types `Document(svg, enc)` actually has.
- `Svg is Aggregable by Text` stays for callers that want just the
  payload and don't care about the encoding (`text.read[Svg]: Svg`).
- `SvgDoc.scala` is deleted; `SvgDoc` removed from the soundness export.

Call-site changes:
  Before: `SvgDoc(svg, enc"UTF-8").xml.show`
  After:  `Document(svg, enc"UTF-8").show`

  Before: `text.read[SvgDoc]`
  After:  `text.load[Svg]: Document[Svg]`

The 67 existing tests carry over (renamed `SvgDoc → Document[Svg]`),
plus a new test asserting that `metadata.name` round-trips through
`load`. `mill savagery.test.run` ⇒ 68 passed / 0 failed.
